client can transmit content which is not in a local file
client and server can both be configured without referencing the global
As2::Config. This allowed the creation of full end-to-end tests using both the
client & server code, and will also be useful when the time comes to rotate out
an expiring certificate. (we can determine which configurations to use at
runtime, and potentially try multiple certificates in sequence if necessary.)

have verified we can correctly send & receive files with mendelson server.