Message: raise if given an invalid MIC algorithm
in expected usage, we'll do validation before passing algorithm name to Message.
so we should only be using default if client does not specify an algorithm.

if client specified an unknown algorithm, & we instead defaulted to sha256, that's
certain to result in a MIC mismatch & a failed transmission - but the silent
fallback will make the issue harder to troubleshoot. so better to raise & solve
the issue at the source.