construct a complete response even if signature verifiation fails
current code raises an exception when `read_smime` receives non-base64 content,
meaning we never populate the disposition field or verify the MIC.

this approach is a workaround. a binary signature will always fail to verify.
but at least we will return a fully-populated MDN and allow our user to decide
whether to consider the transmission a success or not.